Remove dead workspaces
`coverage-map`, `coverage-point-calculator`, and `hex-assignments` are
no longer needed under HIP-149.
