--- coverage_map/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "coverage-map"
-version = "0.1.0"
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-chrono = { workspace = true }
-hex-assignments = { path = "../hex_assignments" }
-hextree = { workspace = true }
