--- coverage_map/src/outdoor.rs
@@ -1,181 +1,0 @@
-use std::{
-    cmp::Ordering,
-    collections::{hash_map::Entry, BinaryHeap, HashMap},
-};
-
-use chrono::{DateTime, Utc};
-use hex_assignments::assignment::HexAssignments;
-use hextree::Cell;
-
-use crate::{CoverageObject, RankedCoverage, SignalLevel, UnrankedCoverage};
-
-/// Data structure for storing outdoor radios ranked by their coverage level
-pub type OutdoorCellTree = HashMap<Cell, BinaryHeap<OutdoorCoverageLevel>>;
-
-#[derive(Eq, Debug, Clone)]
-pub struct OutdoorCoverageLevel {
-    hotspot_key: Vec<u8>,
-    seniority_timestamp: DateTime<Utc>,
-    signal_power: i32,
-    signal_level: SignalLevel,
-    assignments: HexAssignments,
-}
-
-impl PartialEq for OutdoorCoverageLevel {
-    fn eq(&self, other: &Self) -> bool {
-        self.signal_power == other.signal_power
-            && self.seniority_timestamp == other.seniority_timestamp
-    }
-}
-
-impl PartialOrd for OutdoorCoverageLevel {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for OutdoorCoverageLevel {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.signal_power
-            .cmp(&other.signal_power)
-            .reverse()
-            .then_with(|| self.seniority_timestamp.cmp(&other.seniority_timestamp))
-    }
-}
-
-pub fn insert_outdoor_coverage_object(
-    outdoor: &mut OutdoorCellTree,
-    coverage_object: CoverageObject,
-) {
-    for hex_coverage in coverage_object.coverage.into_iter() {
-        insert_outdoor_coverage(
-            outdoor,
-            &coverage_object.hotspot_key,
-            coverage_object.seniority_timestamp,
-            hex_coverage,
-        );
-    }
-}
-
-pub fn insert_outdoor_coverage(
-    outdoor: &mut OutdoorCellTree,
-    hotspot: &[u8],
-    seniority_timestamp: DateTime<Utc>,
-    hex_coverage: UnrankedCoverage,
-) {
-    outdoor
-        .entry(hex_coverage.location)
-        .or_default()
-        .push(OutdoorCoverageLevel {
-            hotspot_key: hotspot.to_vec(),
-            seniority_timestamp,
-            signal_level: hex_coverage.signal_level,
-            signal_power: hex_coverage.signal_power,
-            assignments: hex_coverage.assignments,
-        });
-}
-
-pub fn clone_outdoor_coverage_into_submap(
-    submap: &mut OutdoorCellTree,
-    from: &OutdoorCellTree,
-    coverage_obj: &CoverageObject,
-) {
-    for coverage in &coverage_obj.coverage {
-        if let Entry::Vacant(e) = submap.entry(coverage.location) {
-            if let Some(old_coverage_data) = from.get(&coverage.location) {
-                e.insert(old_coverage_data.clone());
-            }
-        }
-    }
-}
-
-pub fn into_outdoor_coverage_map(outdoor: OutdoorCellTree) -> impl Iterator<Item = RankedCoverage> {
-    outdoor.into_iter().flat_map(move |(hex, radios)| {
-        radios
-            .into_sorted_vec()
-            .into_iter()
-            .enumerate()
-            .map(move |(rank, cov)| RankedCoverage {
-                hex,
-                rank: rank + 1,
-                hotspot_key: cov.hotspot_key,
-                assignments: cov.assignments,
-                signal_level: cov.signal_level,
-            })
-    })
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::*;
-    use chrono::NaiveDate;
-    use hex_assignments::Assignment;
-    use hextree::Cell;
-
-    #[test]
-    fn ensure_outdoor_radios_ranked_by_power() {
-        let mut outdoor_coverage = OutdoorCellTree::default();
-        for cov_obj in vec![
-            outdoor_wifi_coverage("1", -946, date(2022, 8, 1)),
-            outdoor_wifi_coverage("2", -936, date(2022, 12, 5)),
-            outdoor_wifi_coverage("3", -887, date(2022, 12, 2)),
-            outdoor_wifi_coverage("4", -887, date(2022, 12, 1)),
-            outdoor_wifi_coverage("5", -773, date(2023, 5, 1)),
-        ]
-        .into_iter()
-        {
-            insert_outdoor_coverage_object(&mut outdoor_coverage, cov_obj);
-        }
-        let ranked: HashMap<_, _> = into_outdoor_coverage_map(outdoor_coverage)
-            .map(|x| {
-                (
-                    std::str::from_utf8(&x.hotspot_key.clone())
-                        .unwrap()
-                        .to_string(),
-                    x,
-                )
-            })
-            .collect();
-        assert_eq!(ranked.get("5").unwrap().rank, 1);
-        assert_eq!(ranked.get("4").unwrap().rank, 2);
-        assert_eq!(ranked.get("3").unwrap().rank, 3);
-        assert_eq!(ranked.get("1").unwrap().rank, 5);
-        assert_eq!(ranked.get("2").unwrap().rank, 4);
-    }
-
-    fn hex_assignments_mock() -> HexAssignments {
-        HexAssignments {
-            footfall: Assignment::A,
-            urbanized: Assignment::A,
-            landtype: Assignment::A,
-            service_provider_override: Assignment::C,
-        }
-    }
-
-    fn date(year: i32, month: u32, day: u32) -> DateTime<Utc> {
-        NaiveDate::from_ymd_opt(year, month, day)
-            .unwrap()
-            .and_hms_opt(0, 0, 0)
-            .unwrap()
-            .and_utc()
-    }
-
-    fn outdoor_wifi_coverage(
-        hotspot_key: &str,
-        signal_power: i32,
-        seniority_timestamp: DateTime<Utc>,
-    ) -> CoverageObject {
-        CoverageObject {
-            indoor: false,
-            hotspot_key: hotspot_key.as_bytes().to_vec(),
-            seniority_timestamp,
-            coverage: vec![UnrankedCoverage {
-                location: Cell::from_raw(0x8a1fb46622dffff).expect("valid h3 cell"),
-                signal_power,
-                signal_level: SignalLevel::High,
-                assignments: hex_assignments_mock(),
-            }],
-        }
-    }
-}
