--- coverage_point_calculator/src/lib.rs
@@ -1,740 +1,0 @@
-//!
-//! Many changes to the rewards algorithm are contained in and across many HIPs.
-//! The blog post [MOBILE Proof of Coverage][mobile-poc-blog] contains a more
-//! thorough explanation of many of them. It is not exhaustive, but a great
-//! place to start.
-//!
-//! ## Important Fields
-//! - [CoveredHex::points]
-//!   - [HIP-74][modeled-coverage]
-//!   - reduced cbrs radio coverage points [HIP-113][cbrs-experimental]
-//!
-//! - [CoveredHex::assignment_multiplier]
-//!   - [HIP-103][oracle-boosting]
-//!   - [HIP-134][carrier-offload]
-//!     - serving >25 unique connection increase oracle boosting to 1x
-//!   - [HRP-20250409][urban-area-adjustment]
-//!     - all footfall C hexes now have a 0.03 multiplier
-//!
-//! - [CoveredHex::rank]
-//!   - [HIP-105][hex-limits]
-//!
-//! - [CoveragePoints::location_trust_multiplier]
-//!   - [HIP-98][qos-score]
-//!   - states 30m requirement for boosted hexes [HIP-107][prevent-gaming]
-//!   - increase Boosted hex restriction, 30m -> 50m [Pull Request][boosted-hex-restriction]
-//!   - Maximum Asserted Distance Difference [HIP-119][location-gaming]
-//!
-//! - [CoveragePoints::speedtest_multiplier]
-//!   - [HIP-74][modeled-coverage]
-//!   - added "Good" speedtest tier [HIP-98][qos-score]
-//!     - latency is explicitly under limit in HIP <https://github.com/helium/oracles/pull/737>
-//!
-//! ## Notable Conditions:
-//! - [LocationTrust]
-//!   - The average distance to asserted must be <=50m to be eligible for boosted rewards.
-//!   - CBRS Radio's location is always trusted because of GPS.
-//!
-//! - [Speedtest]
-//!   - The latest 6 speedtests will be used.
-//!   - There must be more than 2 speedtests.
-//!
-//! - [SPBoostedRewardEligibility]
-//!   - Radio must serve >25 unique connections on a rolling 7-day window [HIP-140][sp-boost-qualifiers]
-//!   - (deprecated) Service Provider can invalidate boosted rewards of a hotspot [HIP-125][provider-banning]
-//!
-//! - [OracleBoostingStatus]
-//!   - Eligible: Radio is eligible for normal oracle boosting multipliers
-//!   - Qualified: Radio serves >25 unique connections, automatic oracle boosting multiplier of 1x
-//!
-//! ## Deprecated
-//! Service Provider hex boosting ([HIP-84][provider-boosting]) has been removed;
-//! the following no longer affects rewards and is kept for historical reference:
-//! - `CoveredHex::boosted_multiplier` (field removed)
-//!   - Wifi Location trust score >0.75 for boosted hex eligibility [HIP-93][wifi-aps]
-//! - Provider boosted hexes raised the Oracle Assignment multiplier to 1x [HIP-103][oracle-boosting]
-//! - When a Radio was not eligible for boosted rewards, boost values were removed before calculations.
-//!
-//! [modeled-coverage]:        https://github.com/helium/HIP/blob/main/0074-mobile-poc-modeled-coverage-rewards.md#outdoor-radios
-//! [provider-boosting]:       https://github.com/helium/HIP/blob/main/0084-service-provider-hex-boosting.md
-//! [wifi-aps]:                https://github.com/helium/HIP/blob/main/0093-addition-of-wifi-aps-to-mobile-subdao.md
-//! [qos-score]:               https://github.com/helium/HIP/blob/main/0098-mobile-subdao-quality-of-service-requirements.md
-//! [oracle-boosting]:         https://github.com/helium/HIP/blob/main/0103-oracle-hex-boosting.md
-//! [hex-limits]:              https://github.com/helium/HIP/blob/main/0105-modification-of-mobile-subdao-hex-limits.md
-//! [prevent-gaming]:          https://github.com/helium/HIP/blob/main/0107-preventing-gaming-within-the-mobile-network.md
-//! [cbrs-experimental]:       https://github.com/helium/HIP/blob/main/0113-reward-cbrs-as-experimental.md
-//! [mobile-poc-blog]:         https://docs.helium.com/mobile/proof-of-coverage
-//! [boosted-hex-restriction]: https://github.com/helium/oracles/pull/808
-//! [location-gaming]:         https://github.com/helium/HIP/blob/main/0119-closing-gaming-loopholes-within-the-mobile-network.md
-//! [provider-banning]:        https://github.com/helium/HIP/blob/main/0125-temporary-anti-gaming-measures-for-boosted-hexes.md
-//! [anti-gaming]:             https://github.com/helium/HIP/blob/main/0131-bridging-gap-between-verification-mappers-and-anti-gaming-measures.md
-//! [carrier-offload]:         https://github.com/helium/HIP/blob/main/0134-reward-mobile-carrier-offload-hotspots.md
-//! [sp-boost-qualifiers]:     https://github.com/helium/HIP/blob/main/0140-adjust-service-provider-boost-qualifiers.md
-//! [urban-area-adjustment]:   https://github.com/helium/helium-release-proposals/blob/main/releases/20250409-core-devs.md#1-hip-103-urban-area-multiplier-adjustment
-//!
-pub use crate::{
-    hexes::{CoveredHex, HexPoints},
-    location::{asserted_distance_to_trust_multiplier, LocationTrust},
-    service_provider_boosting::SPBoostedRewardEligibility,
-    speedtest::{BytesPs, Speedtest, SpeedtestTier},
-};
-use coverage_map::SignalLevel;
-use rust_decimal::Decimal;
-use rust_decimal_macros::dec;
-use service_provider_boosting::{MAX_AVERAGE_DISTANCE, MIN_WIFI_TRUST_MULTIPLIER};
-
-mod hexes;
-pub mod location;
-mod service_provider_boosting;
-pub mod speedtest;
-
-pub type Result<T = ()> = std::result::Result<T, Error>;
-
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    #[error("signal level {0:?} not allowed for {1:?}")]
-    InvalidSignalLevel(SignalLevel, RadioType),
-    #[error("Input array is empty")]
-    ArrayIsEmpty,
-}
-
-/// Output of calculating coverage points for a Radio.
-///
-/// The data in this struct may be different from the input data, but
-/// it contains the values used for calculating coverage points.
-///
-/// - If more than the allowed speedtests were provided, only the speedtests
-///   considered are included here.
-///
-/// #### Terminology
-///
-/// *Points*
-///
-/// The value provided from covering hexes. These include hex related modifiers.
-/// - Rank
-/// - Oracle boosting.
-///
-/// *Multipliers*
-///
-/// Values relating to a radio that modify it's points.
-/// - Location Trust
-/// - Speedtest
-///
-/// *Shares*
-///
-/// The result of multiplying Points and Multipliers together.
-///
-#[derive(Debug, Clone)]
-pub struct CoveragePoints {
-    /// Breakdown of coverage points by source
-    pub coverage_points: HexPoints,
-    /// Location Trust Multiplier, maximum of 1
-    ///
-    /// Coverage trust of a Radio
-    pub location_trust_multiplier: Decimal,
-    /// Speedtest Multiplier, maximum of 1
-    ///
-    /// Backhaul of a Radio
-    pub speedtest_multiplier: Decimal,
-    /// Input Radio Type
-    pub radio_type: RadioType,
-    /// Input SPBoostedRewardEligibility
-    pub service_provider_boosted_reward_eligibility: SPBoostedRewardEligibility,
-    /// Derived Eligibility for Service Provider Boosted Hex Rewards
-    pub sp_boosted_hex_eligibility: SpBoostedHexStatus,
-    /// Derived Eligibility for Oracle Boosted Hex Rewards
-    pub oracle_boosted_hex_eligibility: OracleBoostingStatus,
-    /// Speedtests used in calculation
-    pub speedtests: Vec<Speedtest>,
-    /// Location Trust Scores used in calculation
-    pub location_trust_scores: Vec<LocationTrust>,
-    /// Covered Hexes used in calculation
-    pub covered_hexes: Vec<CoveredHex>,
-    /// Average speedtest used in calculation
-    pub speedtest_avg: Speedtest,
-}
-
-impl CoveragePoints {
-    pub fn new(
-        radio_type: RadioType,
-        service_provider_boosted_reward_eligibility: SPBoostedRewardEligibility,
-        speedtests: Vec<Speedtest>,
-        location_trust_scores: Vec<LocationTrust>,
-        ranked_coverage: Vec<coverage_map::RankedCoverage>,
-        oracle_boost_status: OracleBoostingStatus,
-    ) -> Result<CoveragePoints> {
-        let location_trust_multiplier = location::multiplier(&location_trust_scores)?;
-
-        let sp_boost_eligibility = SpBoostedHexStatus::new(
-            location_trust_multiplier,
-            &location_trust_scores,
-            service_provider_boosted_reward_eligibility,
-        )?;
-
-        let covered_hexes =
-            hexes::clean_covered_hexes(radio_type, ranked_coverage, oracle_boost_status)?;
-
-        let hex_coverage_points = hexes::calculated_coverage_points(&covered_hexes);
-
-        let speedtests = speedtest::clean_speedtests(speedtests);
-        let (speedtest_multiplier, speedtest_avg) = speedtest::multiplier(&speedtests);
-
-        Ok(CoveragePoints {
-            coverage_points: hex_coverage_points,
-            location_trust_multiplier,
-            speedtest_multiplier,
-            speedtest_avg,
-            radio_type,
-            service_provider_boosted_reward_eligibility,
-            sp_boosted_hex_eligibility: sp_boost_eligibility,
-            oracle_boosted_hex_eligibility: oracle_boost_status,
-            speedtests,
-            location_trust_scores,
-            covered_hexes,
-        })
-    }
-
-    /// Accumulated points related only to coverage.
-    /// (Hex * Rank * Assignment) * Location Trust
-    /// Used for reporting.
-    ///
-    /// NOTE:
-    /// Coverage Points includes Location Trust multiplier. In a future version,
-    /// coverage points will refer only to points received by covering a hex,
-    /// and multipliers like location trust will be applied later to reach a
-    /// value referred to as "shares".
-    ///
-    /// Ref:
-    /// <https://github.com/helium/proto/blob/master/src/service/poc_mobile.proto>
-    /// `message radio_reward`
-    pub fn coverage_points_v1(&self) -> Decimal {
-        self.coverage_points.base * self.location_trust_multiplier
-    }
-
-    /// Accumulated points related to entire radio.
-    /// coverage points * speedtest
-    /// Used in calculating rewards
-    pub fn total_shares(&self) -> Decimal {
-        self.total_base_shares()
-    }
-
-    /// Useful for grabbing only base points when calculating reward shares
-    pub fn total_base_shares(&self) -> Decimal {
-        self.coverage_points.base * self.speedtest_multiplier * self.location_trust_multiplier
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum OracleBoostingStatus {
-    Eligible,
-    Qualified,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SpBoostedHexStatus {
-    Eligible,
-    WifiLocationScoreBelowThreshold(Decimal),
-    AverageAssertedDistanceOverLimit(Decimal),
-    NotEnoughConnections,
-}
-
-impl SpBoostedHexStatus {
-    fn new(
-        location_trust_multiplier: Decimal,
-        location_trust_scores: &[LocationTrust],
-        service_provider_boosted_reward_eligibility: SPBoostedRewardEligibility,
-    ) -> Result<Self> {
-        match service_provider_boosted_reward_eligibility {
-            // hip-140: radio must have enough unique connections
-            SPBoostedRewardEligibility::NotEnoughConnections => Ok(Self::NotEnoughConnections),
-            SPBoostedRewardEligibility::Eligible => {
-                // hip-93: if radio is wifi & location_trust score multiplier < 0.75, no boosting
-                if location_trust_multiplier < MIN_WIFI_TRUST_MULTIPLIER {
-                    return Ok(Self::WifiLocationScoreBelowThreshold(
-                        location_trust_multiplier,
-                    ));
-                }
-
-                // hip-119: if the average distance to asserted is beyond 50m, no boosting
-                let average_distance = location::average_distance(location_trust_scores)?;
-                if average_distance > MAX_AVERAGE_DISTANCE {
-                    return Ok(Self::AverageAssertedDistanceOverLimit(average_distance));
-                }
-
-                Ok(Self::Eligible)
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum RadioType {
-    IndoorWifi,
-    OutdoorWifi,
-}
-
-impl RadioType {
-    fn base_coverage_points(&self, signal_level: &SignalLevel) -> Result<Decimal> {
-        let mult = match self {
-            RadioType::IndoorWifi => match signal_level {
-                SignalLevel::High => dec!(400),
-                SignalLevel::Low => dec!(100),
-                other => return Err(Error::InvalidSignalLevel(*other, *self)),
-            },
-            RadioType::OutdoorWifi => match signal_level {
-                SignalLevel::High => dec!(120),
-                SignalLevel::Medium => dec!(60),
-                SignalLevel::Low => dec!(0),
-                SignalLevel::None => dec!(0),
-            },
-        };
-        Ok(mult)
-    }
-
-    fn rank_multiplier(&self, rank: usize) -> Decimal {
-        match (self, rank) {
-            // Indoor wifi
-            (RadioType::IndoorWifi, 1) => dec!(1),
-            // Outdoor Wifi
-            (RadioType::OutdoorWifi, 1) => dec!(1),
-            (RadioType::OutdoorWifi, 2) => dec!(0.5),
-            (RadioType::OutdoorWifi, 3) => dec!(0.25),
-            // Radios outside acceptable rank in a hex do not get points for that hex.
-            _ => dec!(0),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-
-    use rstest::rstest;
-    use speedtest::SpeedtestTier;
-
-    use super::*;
-    use chrono::Utc;
-    use coverage_map::RankedCoverage;
-    use hex_assignments::{assignment::HexAssignments, Assignment};
-    use rust_decimal_macros::dec;
-
-    #[test]
-    fn service_provider_override_assignment_overrides_other_assignments() {
-        let wifi = CoveragePoints::new(
-            RadioType::IndoorWifi,
-            SPBoostedRewardEligibility::Eligible,
-            speedtest_maximum(),
-            location_trust_maximum(),
-            vec![RankedCoverage {
-                hotspot_key: pubkey(),
-                hex: hex_location(),
-                rank: 1,
-                signal_level: SignalLevel::High,
-                assignments: assignments_from(Assignment::C, true),
-            }],
-            OracleBoostingStatus::Eligible,
-        )
-        .unwrap();
-
-        // Despite the worst possible oracle assignment, the service provider
-        // override grants the maximum assignment multiplier.
-        assert_eq!(dec!(400), wifi.coverage_points_v1());
-    }
-
-    #[test]
-    fn speedtests_effect_reward_shares() {
-        let calculate_indoor_wifi = |speedtests: Vec<Speedtest>| {
-            CoveragePoints::new(
-                RadioType::IndoorWifi,
-                SPBoostedRewardEligibility::Eligible,
-                speedtests,
-                location_trust_maximum(),
-                vec![RankedCoverage {
-                    hotspot_key: pubkey(),
-                    hex: hex_location(),
-                    rank: 1,
-                    signal_level: SignalLevel::High,
-                    assignments: assignments_maximum_no_sp_override(),
-                }],
-                OracleBoostingStatus::Eligible,
-            )
-            .expect("indoor wifi with speedtests")
-        };
-
-        let base_coverage_points = RadioType::IndoorWifi
-            .base_coverage_points(&SignalLevel::High)
-            .unwrap();
-
-        let indoor_wifi = calculate_indoor_wifi(speedtest_maximum());
-        assert_eq!(
-            base_coverage_points * SpeedtestTier::Good.multiplier(),
-            indoor_wifi.total_shares()
-        );
-
-        let indoor_wifi = calculate_indoor_wifi(vec![
-            speedtest_with_download(BytesPs::mbps(88)),
-            speedtest_with_download(BytesPs::mbps(88)),
-        ]);
-        assert_eq!(
-            base_coverage_points * SpeedtestTier::Acceptable.multiplier(),
-            indoor_wifi.total_shares()
-        );
-
-        let indoor_wifi = calculate_indoor_wifi(vec![
-            speedtest_with_download(BytesPs::mbps(62)),
-            speedtest_with_download(BytesPs::mbps(62)),
-        ]);
-        assert_eq!(
-            base_coverage_points * SpeedtestTier::Degraded.multiplier(),
-            indoor_wifi.total_shares()
-        );
-
-        let indoor_wifi = calculate_indoor_wifi(vec![
-            speedtest_with_download(BytesPs::mbps(42)),
-            speedtest_with_download(BytesPs::mbps(42)),
-        ]);
-        assert_eq!(
-            base_coverage_points * SpeedtestTier::Poor.multiplier(),
-            indoor_wifi.total_shares()
-        );
-
-        let indoor_wifi = calculate_indoor_wifi(vec![
-            speedtest_with_download(BytesPs::mbps(25)),
-            speedtest_with_download(BytesPs::mbps(25)),
-        ]);
-        assert_eq!(
-            base_coverage_points * SpeedtestTier::Fail.multiplier(),
-            indoor_wifi.total_shares()
-        );
-    }
-
-    #[test]
-    fn oracle_boosting_assignments_apply_per_hex() {
-        fn ranked_coverage(
-            footfall: Assignment,
-            landtype: Assignment,
-            urbanized: Assignment,
-            service_provider_override: Assignment,
-        ) -> RankedCoverage {
-            RankedCoverage {
-                hotspot_key: pubkey(),
-                hex: hex_location(),
-                rank: 1,
-                signal_level: SignalLevel::High,
-                assignments: HexAssignments {
-                    footfall,
-                    landtype,
-                    urbanized,
-                    service_provider_override,
-                },
-            }
-        }
-
-        use Assignment::*;
-        let outdoor_wifi = CoveragePoints::new(
-            RadioType::OutdoorWifi,
-            SPBoostedRewardEligibility::Eligible,
-            speedtest_maximum(),
-            location_trust_maximum(),
-            vec![
-                // yellow - POI ≥ 1 Urbanized, no SP override
-                ranked_coverage(A, A, A, C), // 120
-                ranked_coverage(A, B, A, C), // 120
-                ranked_coverage(A, C, A, C), // 120
-                // 360
-                // orange - POI ≥ 1 Not Urbanized, no SP override
-                ranked_coverage(A, A, B, C), // 120
-                ranked_coverage(A, B, B, C), // 120
-                ranked_coverage(A, C, B, C), // 120
-                // 360
-                // light green - Point of Interest Urbanized, no SP override
-                ranked_coverage(B, A, A, C), // 84
-                ranked_coverage(B, B, A, C), // 84
-                ranked_coverage(B, C, A, C), // 84
-                ranked_coverage(B, C, A, C), // 84
-                ranked_coverage(B, C, A, C), // 84
-                // 420
-                // dark green - Point of Interest Not Urbanized, no SP override
-                ranked_coverage(B, A, B, C), // 60
-                ranked_coverage(B, B, B, C), // 60
-                ranked_coverage(B, C, B, C), // 60
-                // 180
-                // HRP-20250409 - footfall C
-                ranked_coverage(C, A, A, C), // 3.60
-                ranked_coverage(C, B, A, C), // 3.60
-                ranked_coverage(C, C, A, C), // 3.60
-                ranked_coverage(C, A, B, C), // 3.60
-                ranked_coverage(C, B, B, C), // 3.60
-                ranked_coverage(C, C, B, C), // 3.60
-                // 21.60
-                // gray - Outside of USA, no SP override
-                ranked_coverage(A, A, C, C), // 0
-                ranked_coverage(A, B, C, C), // 0
-                ranked_coverage(A, C, C, C), // 0
-                ranked_coverage(B, A, C, C), // 0
-                ranked_coverage(B, B, C, C), // 0
-                ranked_coverage(B, C, C, C), // 0
-                ranked_coverage(C, A, C, C), // 0
-                ranked_coverage(C, B, C, C), // 0
-                ranked_coverage(C, C, C, C), // 0
-            ],
-            OracleBoostingStatus::Eligible,
-        )
-        .expect("outdoor wifi");
-
-        // 360 + 360 + 420 + 180 + 21.60 = 1341.60
-        assert_eq!(dec!(1341.60), outdoor_wifi.coverage_points_v1());
-    }
-
-    #[rstest]
-    #[case(RadioType::OutdoorWifi, 1, dec!(120))]
-    #[case(RadioType::OutdoorWifi, 2, dec!(60))]
-    #[case(RadioType::OutdoorWifi, 3, dec!(30))]
-    #[case(RadioType::OutdoorWifi, 42, dec!(0))]
-    fn outdoor_radios_consider_top_3_ranked_hexes(
-        #[case] radio_type: RadioType,
-        #[case] rank: usize,
-        #[case] expected_points: Decimal,
-    ) {
-        let outdoor_wifi = CoveragePoints::new(
-            radio_type,
-            SPBoostedRewardEligibility::Eligible,
-            speedtest_maximum(),
-            location_trust_maximum(),
-            vec![RankedCoverage {
-                hotspot_key: pubkey(),
-                hex: hex_location(),
-                rank,
-                signal_level: SignalLevel::High,
-                assignments: assignments_maximum_no_sp_override(),
-            }],
-            OracleBoostingStatus::Eligible,
-        )
-        .expect("outdoor wifi");
-
-        assert_eq!(expected_points, outdoor_wifi.coverage_points_v1());
-    }
-
-    #[rstest]
-    #[case(RadioType::IndoorWifi, 1, dec!(400))]
-    #[case(RadioType::IndoorWifi, 2, dec!(0))]
-    #[case(RadioType::IndoorWifi, 42, dec!(0))]
-    fn indoor_radios_only_consider_first_ranked_hexes(
-        #[case] radio_type: RadioType,
-        #[case] rank: usize,
-        #[case] expected_points: Decimal,
-    ) {
-        let indoor_wifi = CoveragePoints::new(
-            radio_type,
-            SPBoostedRewardEligibility::Eligible,
-            speedtest_maximum(),
-            location_trust_maximum(),
-            vec![
-                RankedCoverage {
-                    hotspot_key: pubkey(),
-                    hex: hex_location(),
-                    rank,
-                    signal_level: SignalLevel::High,
-                    assignments: assignments_maximum_no_sp_override(),
-                },
-                RankedCoverage {
-                    hotspot_key: pubkey(),
-                    hex: hex_location(),
-                    rank: 2,
-                    signal_level: SignalLevel::High,
-                    assignments: assignments_maximum_no_sp_override(),
-                },
-                RankedCoverage {
-                    hotspot_key: pubkey(),
-                    hex: hex_location(),
-                    rank: 42,
-                    signal_level: SignalLevel::High,
-                    assignments: assignments_maximum_no_sp_override(),
-                },
-            ],
-            OracleBoostingStatus::Eligible,
-        )
-        .expect("indoor wifi");
-
-        assert_eq!(expected_points, indoor_wifi.coverage_points_v1());
-    }
-
-    #[test]
-    fn location_trust_score_multiplier() {
-        // Location scores are averaged together
-        let indoor_wifi = CoveragePoints::new(
-            RadioType::IndoorWifi,
-            SPBoostedRewardEligibility::Eligible,
-            speedtest_maximum(),
-            location_trust_with_scores(&[dec!(0.1), dec!(0.2), dec!(0.3), dec!(0.4)]),
-            vec![RankedCoverage {
-                hotspot_key: pubkey(),
-                hex: hex_location(),
-                rank: 1,
-                signal_level: SignalLevel::High,
-                assignments: assignments_maximum_no_sp_override(),
-            }],
-            OracleBoostingStatus::Eligible,
-        )
-        .expect("indoor wifi");
-
-        // Location trust scores is 1/4
-        // (0.1 + 0.2 + 0.3 + 0.4) / 4
-        assert_eq!(dec!(100), indoor_wifi.coverage_points_v1());
-    }
-
-    #[rstest]
-    #[case(SignalLevel::High, dec!(120))]
-    #[case(SignalLevel::Medium, dec!(60))]
-    #[case(SignalLevel::Low, dec!(0))]
-    #[case(SignalLevel::None, dec!(0))]
-    fn outdoor_wifi_base_coverage_points(
-        #[case] signal_level: SignalLevel,
-        #[case] expected: Decimal,
-    ) {
-        let outdoor_wifi = CoveragePoints::new(
-            RadioType::OutdoorWifi,
-            SPBoostedRewardEligibility::Eligible,
-            speedtest_maximum(),
-            location_trust_maximum(),
-            vec![RankedCoverage {
-                hotspot_key: pubkey(),
-                hex: hex_location(),
-                rank: 1,
-                signal_level,
-                assignments: assignments_maximum_no_sp_override(),
-            }],
-            OracleBoostingStatus::Eligible,
-        )
-        .expect("outdoor wifi");
-
-        assert_eq!(expected, outdoor_wifi.coverage_points_v1());
-    }
-
-    #[rstest]
-    #[case(SignalLevel::High, dec!(400))]
-    #[case(SignalLevel::Low, dec!(100))]
-    fn indoor_wifi_base_coverage_points(
-        #[case] signal_level: SignalLevel,
-        #[case] expected: Decimal,
-    ) {
-        let indoor_wifi = CoveragePoints::new(
-            RadioType::IndoorWifi,
-            SPBoostedRewardEligibility::Eligible,
-            speedtest_maximum(),
-            location_trust_maximum(),
-            vec![RankedCoverage {
-                hotspot_key: pubkey(),
-                hex: hex_location(),
-                rank: 1,
-                signal_level,
-                assignments: assignments_maximum_no_sp_override(),
-            }],
-            OracleBoostingStatus::Eligible,
-        )
-        .expect("indoor wifi");
-
-        assert_eq!(expected, indoor_wifi.coverage_points_v1());
-    }
-
-    #[test]
-    fn wifi_with_bad_location_boosted_hex_status_prioritizes_service_provider_statuses() {
-        let bad_location = vec![LocationTrust {
-            meters_to_asserted: 100,
-            trust_score: dec!(0.0),
-        }];
-
-        let wifi_bad_trust_score = |sp_status: SPBoostedRewardEligibility| {
-            SpBoostedHexStatus::new(
-                location::multiplier(&bad_location).unwrap(),
-                &bad_location,
-                sp_status,
-            )
-            .unwrap()
-        };
-
-        assert_eq!(
-            wifi_bad_trust_score(SPBoostedRewardEligibility::Eligible),
-            SpBoostedHexStatus::WifiLocationScoreBelowThreshold(dec!(0)),
-        );
-        assert_eq!(
-            wifi_bad_trust_score(SPBoostedRewardEligibility::NotEnoughConnections),
-            SpBoostedHexStatus::NotEnoughConnections
-        );
-    }
-
-    fn hex_location() -> hextree::Cell {
-        hextree::Cell::from_raw(0x8c2681a3064edff).unwrap()
-    }
-
-    fn assignments_maximum_no_sp_override() -> HexAssignments {
-        HexAssignments {
-            footfall: Assignment::A,
-            landtype: Assignment::A,
-            urbanized: Assignment::A,
-            service_provider_override: Assignment::C,
-        }
-    }
-
-    fn assignments_from(assignment: Assignment, service_provider_override: bool) -> HexAssignments {
-        let service_provider_override_assignment = if service_provider_override {
-            Assignment::A
-        } else {
-            Assignment::C
-        };
-        HexAssignments {
-            footfall: assignment,
-            landtype: assignment,
-            urbanized: assignment,
-            service_provider_override: service_provider_override_assignment,
-        }
-    }
-
-    fn speedtest_maximum() -> Vec<Speedtest> {
-        vec![
-            Speedtest {
-                upload_speed: BytesPs::mbps(15),
-                download_speed: BytesPs::mbps(150),
-                latency_millis: 15,
-                timestamp: Utc::now(),
-            },
-            Speedtest {
-                upload_speed: BytesPs::mbps(15),
-                download_speed: BytesPs::mbps(150),
-                latency_millis: 15,
-                timestamp: Utc::now(),
-            },
-        ]
-    }
-
-    fn speedtest_with_download(download: BytesPs) -> Speedtest {
-        Speedtest {
-            upload_speed: BytesPs::mbps(15),
-            download_speed: download,
-            latency_millis: 15,
-            timestamp: Utc::now(),
-        }
-    }
-
-    fn location_trust_maximum() -> Vec<LocationTrust> {
-        vec![LocationTrust {
-            meters_to_asserted: 1,
-            trust_score: dec!(1.0),
-        }]
-    }
-
-    fn location_trust_with_scores(trust_scores: &[Decimal]) -> Vec<LocationTrust> {
-        trust_scores
-            .to_owned()
-            .iter()
-            .copied()
-            .map(|trust_score| LocationTrust {
-                meters_to_asserted: 1,
-                trust_score,
-            })
-            .collect()
-    }
-
-    fn pubkey() -> Vec<u8> {
-        vec![1]
-    }
-}
