--- hex_assignments/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "hex-assignments"
-version = "0.1.0"
-description = "Hex Assignments"
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-
-[dependencies]
-anyhow = { workspace = true }
-hextree = { workspace = true }
-sqlx = { workspace = true }
-rust_decimal = { workspace = true }
-rust_decimal_macros = { workspace = true }
-helium-proto = { workspace = true }
-async-trait = { workspace = true }
-chrono = { workspace = true }
-derive_builder = { workspace = true }
-tls-init = { path = "../tls_init" }
